More column-generic tests
- roundtrip to server
