+++ ut/Column_ut.cpp
@@ -13,9 +13,12 @@
 #include <clickhouse/base/output.h>
 #include <clickhouse/base/socket.h> // for ipv4-ipv6 platform-specific stuff
 
+#include <clickhouse/client.h>
+
 #include <gtest/gtest.h>
 
 #include "utils.h"
+#include "roundtrip_column.h"
 #include "value_generators.h"
 
 namespace {
@@ -261,4 +264,21 @@ TYPED_TEST(GenericColumnTest, LoadAndSave) {
     }
 
     EXPECT_TRUE(CompareRecursive(*column_A, *column_B));
+}
+
+const auto LocalHostEndpoint = ClientOptions()
+        .SetHost(           getEnvOrDefault("CLICKHOUSE_HOST",     "localhost"))
+        .SetPort(   getEnvOrDefault<size_t>("CLICKHOUSE_PORT",     "9000"))
+        .SetUser(           getEnvOrDefault("CLICKHOUSE_USER",     "default"))
+        .SetPassword(       getEnvOrDefault("CLICKHOUSE_PASSWORD", ""))
+        .SetDefaultDatabase(getEnvOrDefault("CLICKHOUSE_DB",       "default"));
+
+TYPED_TEST(GenericColumnTest, RoundTrip) {
+    auto [column, values] = this->MakeColumnWithValues(100);
+    EXPECT_EQ(values.size(), column->Size());
+
+    clickhouse::Client client(LocalHostEndpoint);
+
+    auto result_typed = RoundtripColumnValues(client, column)->template AsStrict<typename TestFixture::ColumnType>();
+    EXPECT_TRUE(CompareRecursive(*column, *result_typed));
 }
