+++ ut/roundtrip_column.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <clickhouse/columns/column.h>
+
+namespace clickhouse {
+    class Client;
+}
+
+clickhouse::ColumnRef RoundtripColumnValues(clickhouse::Client& client, clickhouse::ColumnRef expected);
